services/resizer/pyproject.toml
[tool.poetry]
name = "sizer"
version = "0.1.0"
description = ""
authors = ["Simon Bumm <mail@codingcyclist.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-lambda-powertools = "^2.25.1"
pillow = "^10.1.0"
moto = "^4.2.12"
pytest = "^7.4.3"

[tool.poetry.group.dev.dependencies]
types-pillow = "^10.1.0.0"
boto3 = "^1.28.73"
mypy-boto3 = "^1.28.73"
mypy-boto3-s3 = "^1.28.55"
boto3-stubs = "^1.28.73"
aws-xray-sdk = "^2.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
